+++ packages/protocol/contracts/L1/TaikoL1.sol
@@ -46,11 +46,19 @@ contract TaikoL1 is EssentialContract, ITaikoL1, ITierProvider, TaikoEvents, Tai
         if (!_inNonReentrant()) revert L1_RECEIVE_DISABLED();
     }
 
-    /// @notice Initializes the rollup.
-    /// @param _addressManager The {AddressManager} address.
+    /// @notice Initializes the contract.
+    /// @param _owner The owner of this contract. msg.sender will be used if this value is zero.
+    /// @param _addressManager The address of the {AddressManager} contract.
     /// @param _genesisBlockHash The block hash of the genesis block.
-    function init(address _addressManager, bytes32 _genesisBlockHash) external initializer {
-        __Essential_init(_addressManager);
+    function init(
+        address _owner,
+        address _addressManager,
+        bytes32 _genesisBlockHash
+    )
+        external
+        initializer
+    {
+        __Essential_init(_owner, _addressManager);
         LibVerifying.init(state, getConfig(), _genesisBlockHash);
     }
 
@@ -70,8 +78,7 @@ contract TaikoL1 is EssentialContract, ITaikoL1, ITierProvider, TaikoEvents, Tai
     {
         TaikoData.Config memory config = getConfig();
 
-        (meta, depositsProcessed) =
-            LibProposing.proposeBlock(state, config, AddressResolver(this), params, txList);
+        (meta, depositsProcessed) = LibProposing.proposeBlock(state, config, this, params, txList);
 
         if (!state.slotB.provingPaused) {
             _verifyBlocks(config, config.maxBlocksToVerifyPerProposal);
@@ -98,8 +105,7 @@ contract TaikoL1 is EssentialContract, ITaikoL1, ITierProvider, TaikoEvents, Tai
 
         TaikoData.Config memory config = getConfig();
 
-        uint8 maxBlocksToVerify =
-            LibProving.proveBlock(state, config, AddressResolver(this), meta, tran, proof);
+        uint8 maxBlocksToVerify = LibProving.proveBlock(state, config, this, meta, tran, proof);
 
         _verifyBlocks(config, maxBlocksToVerify);
     }
@@ -120,11 +126,11 @@ contract TaikoL1 is EssentialContract, ITaikoL1, ITierProvider, TaikoEvents, Tai
     /// @param recipient Address of the recipient for the deposited Ether on
     /// Layer 2.
     function depositEtherToL2(address recipient) external payable nonReentrant whenNotPaused {
-        LibDepositing.depositEtherToL2(state, getConfig(), AddressResolver(this), recipient);
+        LibDepositing.depositEtherToL2(state, getConfig(), this, recipient);
     }
 
     function unpause() public override {
-        OwnerUUPSUpgradable.unpause(); // permission checked inside
+        super.unpause(); // permission checked inside
         state.slotB.lastUnpausedAt = uint64(block.timestamp);
     }
 
@@ -252,7 +258,7 @@ contract TaikoL1 is EssentialContract, ITaikoL1, ITierProvider, TaikoEvents, Tai
         internal
         whenProvingNotPaused
     {
-        LibVerifying.verifyBlocks(state, config, AddressResolver(this), maxBlocksToVerify);
+        LibVerifying.verifyBlocks(state, config, this, maxBlocksToVerify);
     }
 
     function _authorizePause(address)