+++ packages/protocol/contracts/L1/libs/LibUtils.sol
@@ -14,7 +14,7 @@
 
 pragma solidity 0.8.24;
 
-import "../../common/AddressResolver.sol";
+import "../../common/IAddressResolver.sol";
 import "../../signal/ISignalService.sol";
 import "../../signal/LibSignals.sol";
 import "../TaikoData.sol";