+++ packages/protocol/contracts/L1/tiers/DevnetTierProvider.sol
@@ -22,9 +22,10 @@ import "./ITierProvider.sol";
 contract DevnetTierProvider is EssentialContract, ITierProvider {
     uint256[50] private __gap;
 
-    /// @notice Initializes the contract with the provided address manager.
-    function init() external initializer {
-        __Essential_init();
+    /// @notice Initializes the contract.
+    /// @param _owner The owner of this contract. msg.sender will be used if this value is zero.
+    function init(address _owner) external initializer {
+        __Essential_init(_owner);
     }
 
     function getTier(uint16 tierId) public pure override returns (ITierProvider.Tier memory) {