+++ packages/protocol/contracts/signal/SignalService.sol
@@ -67,9 +67,11 @@ contract SignalService is EssentialContract, ISignalService {
         _;
     }
 
-    /// @dev Initializer to be called after being deployed behind a proxy.
-    function init(address _addressManager) external initializer {
-        __Essential_init(_addressManager);
+    /// @notice Initializes the contract.
+    /// @param _owner The owner of this contract. msg.sender will be used if this value is zero.
+    /// @param _addressManager The address of the {AddressManager} contract.
+    function init(address _owner, address _addressManager) external initializer {
+        __Essential_init(_owner, _addressManager);
     }
 
     /// @dev Authorize or deautohrize an address for calling syncChainData